Add an endpoint to retrieve data about a person via their internal id or via their name.
Limitations: You need to get there name as it exist
in the DB exactly right.

Up to casing

A name may not be unique, it may return multiple rows. We fail for now on that case.

The name may not exist we will fail on that.

ID may not exist we will fail on that.

Some of those cases should be 404 errors (not found) rather than 500 (internal unspecfiied errors)